backend/requirements.txt
fastapi
uvicorn
motor
pydantic
tensorflow
spacy
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
pydantic[email]
pydantic-settings
